% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonConnector_UI.R
\name{python_version_UI}
\alias{python_version_UI}
\title{Python Version}
\usage{
python_version_UI(id)
}
\arguments{
\item{id}{the namespace used to call the module server side.}
}
\value{
a text output containing active python version.
}
\description{
Python Version
}
\examples{
python_version_UI("python")
}
