% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonConnector_Server.R
\name{pythonConnector_Server}
\alias{pythonConnector_Server}
\title{Title Python Connector Server}
\usage{
pythonConnector_Server(id, script_path, script_file = NULL)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Python connector implementation through Reticulate package.
configure Python virtualenv and paths in .Rprofile file.
}
