% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonConnector_UI.R
\name{sysinfo_UI}
\alias{sysinfo_UI}
\title{Python Version}
\usage{
sysinfo_UI(id)
}
\arguments{
\item{id}{the namespace used to call the module server side.}
}
\value{
a DT table output containing system info.
}
\description{
Python Version
}
\examples{
sysinfo_UI("python")
}
