% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonConnector_UI.R
\name{which_python_UI}
\alias{which_python_UI}
\title{Which Python}
\usage{
which_python_UI(id)
}
\arguments{
\item{id}{the namespace used to call the module server side.}
}
\value{
a text output with path to the active python.
}
\description{
Which Python
}
\examples{
which_python_UI("python")
}
