% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonConnector_Server.R
\name{pythonConnector_Server}
\alias{pythonConnector_Server}
\title{Title Python Connector Server}
\usage{
pythonConnector_Server(
  id,
  script_path,
  script_file = NULL,
  dependencies = NULL
)
}
\arguments{
\item{id}{server module id}

\item{script_path}{path to the python script files}

\item{script_file}{if provided, only this list of scripts will be loaded}

\item{dependencies}{path where to find the python_requirements.txt file containing the package dependencies for installation}
}
\description{
Python connector implementation through Reticulate package.
configure Python virtualenv and paths in .Rprofile file.
}
